+++ src/test/java/kr/hhplus/be/server/application/payment/PaymentServiceTest.java
@@ -33,7 +33,7 @@ class PaymentServiceTest {
     void initiate_createsPayment() {
         // given
         String orderId = "order-1";
-        BigDecimal amount = BigDecimal.valueOf(10000);
+        long amount = 10000L;
         String method = "BALANCE";
 
         // when
